requirements.txt
asgiref
boto3==1.20.26
botocore==1.23.54
dj-database-url==0.5.0
Django
django-auto-logout
django-ckeditor
django-ckeditor-5==0.1.6
django-crispy-forms==1.12.0
django-dotenv==1.4.2
django-environ==0.9.0
django-filter
django-formset-js-improved==0.5.0.3
django-heroku
django-jazzmin
django-js-asset
django-plaintext-password==0.2.0
django-rest-auth
django-rest-framework==0.1.0
django-social-share==2.2.1
django-static-fontawesome==5.14.0.0
django-storages==1.12.3
django-taggit==3.0.0
django-templated-mail==1.1.1
django-tinymce==3.4.0
django-widget-tweaks==1.4.8
djangorestframework
djangorestframework-simplejwt
djoser
gunicorn==20.0.4
lxml
pandas
Pillow
psycopg2
psycopg2-binary
python-decouple==3.5
requests==2.27.1
requests-oauthlib==1.3.1
rjsmin==1.1.0
s3transfer==0.5.2
sendgrid==6.9.7
uritemplate==4.1.1
urllib3==1.26.9
validate-email==1.3
whitenoise==5.2.0
django-import-export
shortuuid
django-shortuuidfield